navigation: refactor <nav> tag into component Navbar
