+++ src/components/Layout/Layout.component.jsx
@@ -1,47 +1,13 @@
 import React from 'react';
-import { Link, useHistory } from 'react-router-dom';
-import { useAuth } from '../../providers/Auth';
-import SearchForm from '../Search/SearchForm.component';
-
-import './Layout.styles.css';
+import Navbar from '../Nav/Navbar.component';
 
 export default function Layout({ children }) {
-  const history = useHistory();
-  const { authenticated, logout } = useAuth();
-
-  function deAuthenticate(event) {
-    event.preventDefault();
-    logout();
-    history.push('/');
-  }
-
   return (
     <>
       <header>
-        <nav>
-          <div>
-            <ul>
-              <li>
-                <Link to="/">Home</Link>
-              </li>
-              <li>
-                <Link to="/favorites">Favorites</Link>
-              </li>
-            </ul>
-            <SearchForm />
-            {authenticated ? (
-              <>
-                <Link to="/" onClick={deAuthenticate}>
-                  ← logout
-                </Link>
-              </>
-            ) : (
-              <Link to="/login">login →</Link>
-            )}
-          </div>
-        </nav>
+        <Navbar />
       </header>
-      <main className="container">{children}</main>
+      <main className="main">{children}</main>
     </>
   );
 }
